frontend/python_host/requirements.txt
pywebview>=5.1.0
